pyproject.toml
[tool.poetry]
name = "bighead"
version = "0.1.0"
description = "Detects the largest face in an image"
authors = ["Chris de Graaf <chrisadegraaf@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6"
Flask = "^1.1.1"
dlib = "^19.19.0"
numpy = "^1.18.1"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
mypy = "^0.761"
pytest-cov = "^2.8.1"
pytest-black = "^0.3.8"
pytest-flake8 = "^1.0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
